[RNMobile] Embed Block Bottom Sheet to set a link (#31343)
* Add separatorType option to LinkSettings footer

* Create native EmbedBottomSheet component

* Use EmbedBottomSheet in native EmbedPlaceholder

* Add onSetAttributes callback to embed bottom sheet

When the attributes are modified in the LinkSettingsNavigation component the setAttributes callback is triggered. This callback is called once unlike the onClose callback, which is called multiple times and doesn't guarantee that the attributes are up-to-date at the time is triggered.

* Correct isEmbedSheetVisible typo

* Use onDismiss callback to notify embed URL value update

* Add comment to onChangeURL embed placeholder

* Only display the embed sheet on block insertion

* Add native version of onSubmit in embed

* Prevent onDismiss prop overwriting in bottom sheet

I noticed that if a onDismiss prop is passed to the bottom sheet component, the internal function onDismiss is not executed. This behavior looks wrong because this function is intended to actually call the onDismiss prop.

* Show only the embed sheet automatically if empty

* Fix clipboard case for embed sheet

* Omit onDismiss prop from rest object

Co-authored-by: Carlos Garcia <fluiddot@gmail.com>