Remove padding from Home Link block (#33461)
No other link block in core includes padding, yet the Home Link block does. This forces theme developers to override this padding to ensure link consistency. Removing this default padding will improve implementation of the Navigation block. The current workaround it to use a Custom Link that manually directs to the site's homepage.