Fix unwanted additional spaces added around pasted text on Windows  (#33607)
* Remove the whitespace before the fragments that are provided by windows

See https://discuss.prosemirror.net/t/space-added-on-paste/1274/7

* Extract to util

* Avoid creating public APIs for util and improve function naming for clarity

As we're not sure about this fix let's not create a public API via utils. Instead revert to inlining.

Also renaming the function to make it's purpose clearer. It's not a generic normalizing function so let's make it obvious this is Windows specific.

* Make both regexes literal