Fix build hang on Windows 10 (#23589)
On Windows 10, I seem to be unable to build. The build completion rises
to 100%, but the process never finishes. I can build if I manually
set the number of workers in the farm to 4, but not if it's any higher.

It appears that some files are not being counted as "complete". It's
probably files that finish before the "ended" signal is recieved (meaning
everything has been queued.) Moving the increment out of the conditional
fixes this.