[RNMobile] Resolve Text Entry Issues/Errors in Alt Text Settings  (#33975)
* Resolve render error

This commit aims to fix the render error outlined here: https://github.com/WordPress/gutenberg/pull/33057#pullrequestreview-725305368

The "outside" changes that happen when text is changed in the BottomSheetTextControl component should not be called directly from that component. Instead, they're now called within the useEffect hook.

* Replace "value" prop with "defaultValue"

The purpose of this commit is to address the text entry issues that currently exist for the TextInput component, as described here: https://github.com/facebook/react-native/issues/30503

A working workaround propose in that thread is to replace the "value" prop with "defaultValue".

* Simplify component by removing unecessary hooks

The introduction of the "defaultValue" prop in https://github.com/WordPress/gutenberg/pull/33975/commits/f6d9697ab66a05e43d253f853de8ec74f12a73b6 means that it's no longer necessary to keep the value in state. "defaultValue" allows for text to be automatically changed as it's typed. See: https://reactnative.dev/docs/textinput#defaultvalue

We don't need to use "onChangeText" for the purpose of updating the input's text and can instead use it to invoke the "onChange" function.

We can also remove the useState and useEffect hooks, as they're not necessary to update the component's value.