Add editor onboarding tests (#34431)
* Add tooltip display message test

* Add tap screen dismisses tooltip test

* Add keyboard hide dismisses tooltip test

* Simplify tooltip tests

* Fix React Native warning caused by webpack-oriented Jest image mock

The previous image mock was implemented following the Jest documentation
related to webpack. However, the `source` prop for React Native's Image
component does not accept strings. The previous image mock resulted in
the following warning in testing logs.

```
Warning: Failed prop type: Invalid prop `source` supplied to `Image`.
```

To address this, the previous image mock was removed in favor of relying
upon the transform that is provided by the `react-native` Jest preset.
In order to have this transform applied to images within React
Navigation, additional regex was added to `transformIgnorePatterns`.

- https://git.io/JEFo0
- https://bit.ly/3jAqaJL
- https://bit.ly/2YiuDbS
- https://git.io/JEFor
- https://git.io/JEFoK

* Add editor Help section tests

* Replace unnecessary React Hook check

Defining a functioning `remove` method for the `Keyboard` mock is more
appropriate than modifying the application code for testing purposes.
The original mock did not provide a way to remove registered callbacks,
which resulted in stale callbacks updating React state erroneously.

* Remove redundant assertion

A sibling test asserts the presence of this text. The presence of the
back button also provides similar coverage

* Use npm install instead of npm ci

Temporarily change install method to circumvent erroneous CI failures.

* Revert "Use npm install instead of npm ci"

This reverts commit ba00594639fefeee5852380a0b6bf24fffa683f4.