Enable native editor onboarding features globally (#35201)
* Enable new block badge globally

We decided to enable the new block badge for all users.

* Update release notes

* Fix block inserter unit test

Now that block impressions are enabled for all users, we must update the
test expectation to include the `isNew` property set by
`useBlockTypeImpression`. An alternative would be to update this test to
not explicitly check the component prop, but the rendered result which
the user can see.

* Remove editor onboarding bridging methods

The bridging methods are no longer necessary now that we display editor
onboarding features for all users.

* Remove help support articles from release notes

This feature was delivered in a previous release via a beta fix.

* Remove first block editor session flag

* Remove unnecessary trailing comma

* Relocate native-specific fixture setup

Place fixture setup in fixture file rather than within test file.