Multi-entity save: Only set site entity to pending if really saving (#36573)
Currently, the multi-entity save panel is unconditionally attempting to save changes to the site entity, even if there aren't any. This causes the site entity saving state to be set to "pending" at least for a short period of time.

This commit makes it so that we only save the site entity if there are any changes to it.