Removing unused className prop on BlockStyles (#36976)
Removing unused stylesheet for default style picker
Adding a wrapper around the default picker select control so that we can target the component container
Adding a sibling selector to block styles to give the default picker a margin top. We can specifically target the default picker only when it appears after the block styles component.