Prompt to keep or replace Featured Image when replacing Image block media that is set as the Featured Image (#34666)
* Replacing Image block media that is featured prompts user

If an Image block is set as the post's Featured Image, replacing the
Image block's media will prompt the user as to whether they want to
replace the Featured Image as well.

* Remove unused assignment

* Fix replace Featured Image prompt when uploading a new image

When media is uploaded as a replacement for an Image block's media, we
must await the final, non-negative ID before attempting to set the new
media as the Featured Image.

* Revert unnecessary change

* Update change log

* Disallow setting in-progress or failed uploads as featured

In-progress or failed image uploads cannot be set as the featured image
as the media does not exist on the server. Relying upon a negative ID
number did not properly disable in-progress uploads, as negative IDs are
a iOS-only coincidence due to hashing of the integer ID.

These changes rely upon media upload sync events and the presence of an
`image` prop to determine when the upload is completed successfully.

* Refactor multiple status booleans to single enum

This was done to reduce the likelihood of conflicting states. Now the
upload status can only occupy a single state.