Post comments form block show warning wrt to post/page (#38011)
Earlier, we were showing a single hardcorded message irrespective of
post type. Warning message used to write post even in case of WP page.
It would be better if we show post in case of post and page in case of
acutal page. Merging this change will also help user in terms of reading
and understanding warning message.