docgen: Unwrap wrapped selectors when inferring types of JSDoc params (#40236)
In #40025 we ran into an obstacle when transitioning code into TypeScript,
namely that our `docgen` tool is unable to find the associated types for
parameters which are documented in JSDoc comments if the parameters are
for the returned function from a call to some wrapping function.

In this patch we're adding two special cases for selectors that call
`createSelector` and `createRegistrySelector` to allow our `docgen` tool
to analyze those inner functions which represent the actual selector.

Fundamentally we should be asking TypeScript for the inferred types of
the function and its parameters but given that we don't have a current
mechanism to do that this issue remains a blocker for broader TypeScript
work. Because of that we're introducing hard-coded special cases for
these common selector wrappers so that we can unblock the TypeScript
work without introducing a generic compromise with potentially-harmful
side-effects, such as might happen if we were to always return the
first argument of a call expression.