docgen: Replace fixtures with inline parsed code. (#40428)
Previously in the `docgen` package test suite we relied on developers manually creating
a set of files for each test that includes a reference snippet of code, a parsed AST,
and an exports listing. These files weren't linked together in any way.

In this patch we're removing the manual steps and inlining the code snippets inside each
test in the suite. This removes a level of manual work and indirection when maintaining
the tests while leaning more on `docgen`'s `engine` component, which itself does the
actual parsing.